requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-filter==23.3
django-ratelimit==4.1.0
django-environ==0.11.2
Pillow==10.0.1
cryptography==41.0.7
twilio==8.10.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.7
mysqlclient==2.2.0
gunicorn==21.2.0
whitenoise==6.6.0
django-audit-log==0.0.1
django-otp==1.3.0
django-two-factor==1.15.5
clamd==1.0.2
reportlab==4.0.4
xlsxwriter==3.1.9
channels>=4.0.0